add: make canScheduleExactAlarms() to have version compatible add: pull some features from main view model and put in new view model (app view model)
