+++ app/src/main/java/com/joeloewi/croissant/receiver/ResinStatusWidgetProvider.kt
@@ -14,7 +14,7 @@ import android.widget.RemoteViews
 import androidx.work.*
 import com.joeloewi.croissant.R
 import com.joeloewi.croissant.util.goAsync
-import com.joeloewi.croissant.util.isIgnoringBatteryOptimizations
+import com.joeloewi.croissant.util.isIgnoringBatteryOptimizationsCompat
 import com.joeloewi.croissant.util.pendingIntentFlagUpdateCurrent
 import com.joeloewi.croissant.worker.RefreshResinStatusWorker
 import com.joeloewi.domain.usecase.ResinStatusWidgetUseCase
@@ -53,7 +53,7 @@ class ResinStatusWidgetProvider : AppWidgetProvider() {
         ) {
             appWidgetIds?.map { appWidgetId ->
                 async(Dispatchers.IO) {
-                    if (powerManager.isPowerSaveMode && !powerManager.isIgnoringBatteryOptimizations(
+                    if (powerManager.isPowerSaveMode && !powerManager.isIgnoringBatteryOptimizationsCompat(
                             application
                         )
                     ) {
