+++ app/src/main/java/com/joeloewi/croissant/util/AlarmManagerExtensions.kt
@@ -1,0 +1,11 @@
+package com.joeloewi.croissant.util
+
+import android.app.AlarmManager
+import android.os.Build
+
+fun AlarmManager.canScheduleExactAlarmsCompat() =
+    if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.S) {
+        canScheduleExactAlarms()
+    } else {
+        true
+    }
