+++ app/src/main/java/com/joeloewi/croissant/util/PowerManagerExtensions.kt
@@ -4,7 +4,7 @@ import android.app.Application
 import android.os.Build
 import android.os.PowerManager
 
-fun PowerManager.isIgnoringBatteryOptimizations(application: Application) =
+fun PowerManager.isIgnoringBatteryOptimizationsCompat(application: Application) =
     if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
         isIgnoringBatteryOptimizations(application.packageName)
     } else {
